Move QuestionAnswerAdvisor and VectorStoreChatMemoryAdvisor
To be consistent with spring-ai-rag and spring-ai-tool-search-tool,
QuestionAnswerAdvisor and VectorStoreChatMemoryAdvisor have moved:
- From spring-ai-advisors-vector-store module to the spring-ai-vector-store one
- From org.springframework.ai.chat.client.advisor.vectorstore to
  org.springframework.ai.vectorstore.advisor one
