+++ spring-ai-vector-store/src/main/java/org/springframework/ai/vectorstore/advisor/VectorStoreChatMemoryAdvisor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.ai.chat.client.advisor.vectorstore;
+package org.springframework.ai.vectorstore.advisor;
 
 import java.util.ArrayList;
 import java.util.HashMap;
